extracted AllModelsTab, eliminating props
